K3OS: support http proxy and exclude offline images

+++ k3os/install.sh
@@ -160,7 +160,6 @@ do_copy()
         echo "Decompressing container images"
         zstd -d --rm "${root_path}/${offline_image_path}.zst" -o "${root_path}/${offline_image_path}" > /dev/null
     fi
-    echo "Loading images. This may take a few minutes"
     cd ${root_path}
     mkdir lib bin sbin k3os dev proc etc sys
     mount --bind /bin bin
@@ -171,28 +170,31 @@ do_copy()
     mount --bind /etc etc
     mount -r --rbind /lib lib
     mount -r --rbind /sys sys
-    chroot . /bin/bash <<"EOF"
-    # invoke k3s to set up data dir
-    k3s agent --no-flannel &>/dev/null || true
-    # start containerd
-    /var/lib/rancher/k3s/data/current/bin/containerd \
-    -c /var/lib/rancher/k3s/agent/etc/containerd/config.toml \
-    -a /run/k3s/containerd/containerd.sock \
-    --state /run/k3s/containerd \
-    --root /var/lib/rancher/k3s/agent/containerd &>/dev/null &
+    if [ -f "${root_path}/${offline_image_path}" ]; then
+      echo "Loading images. This may take a few minutes"
+      chroot . /bin/bash <<"EOF"
+      # invoke k3s to set up data dir
+      k3s agent --no-flannel &>/dev/null || true
+      # start containerd
+      /var/lib/rancher/k3s/data/current/bin/containerd \
+      -c /var/lib/rancher/k3s/agent/etc/containerd/config.toml \
+      -a /run/k3s/containerd/containerd.sock \
+      --state /run/k3s/containerd \
+      --root /var/lib/rancher/k3s/agent/containerd &>/dev/null &
 
-    #wait for containerd to be ready
-    until ctr --connect-timeout 1s version>/dev/null
-    do
-      sleep 1
-    done
-    # import images
-    ctr -n k8s.io images import /var/lib/rancher/k3s/agent/images/harvester*
-    rm /var/lib/rancher/k3s/agent/images/harvester*
-    # stop containerd
-    pkill containerd
-    exit
+      #wait for containerd to be ready
+      until ctr --connect-timeout 1s version>/dev/null
+      do
+        sleep 1
+      done
+      # import images
+      ctr -n k8s.io images import /var/lib/rancher/k3s/agent/images/harvester*
+      rm /var/lib/rancher/k3s/agent/images/harvester*
+      # stop containerd
+      pkill containerd
+      exit
 EOF
+    fi
     sleep 5
     #cleanup
     umount bin sbin k3os dev proc etc
